+++ packages/article-paragraph/src/styles/index.web.js
@@ -3,9 +3,5 @@ import sharedStylesFactory from "./shared";
 const sharedStyles = sharedStylesFactory();
 
 export default {
-  ...sharedStyles,
-  articleTextElement: {
-    ...sharedStyles.articleTextElement,
-    marginTop: 0
-  }
+  ...sharedStyles
 };