+++ packages/save-and-share-bar/__tests__/shared.base.js
@@ -1,5 +1,6 @@
 import React from "react";
 import TestRenderer from "react-test-renderer";
+import saveApi from "@times-components/save-star-web/mock-save-api";
 import { Clipboard } from "react-native";
 import "./mocks";
 import BarItem from "../src/bar-item";
@@ -8,8 +9,7 @@ import SaveAndShareBar from "../src/save-and-share-bar";
 export default () => {
   describe("save and share bar component", () => {
     const onCopyLink = jest.fn();
-    const onSaveToMyArticles = jest.fn();
-    const articleId = "id-123";
+    const articleId = "96508c84-6611-11e9-adc2-05e1b87efaea";
     const articleUrl = "https://www.thetimes.co.uk/";
     const articleHeadline = "test-headline";
 
@@ -23,22 +23,19 @@ export default () => {
           articleHeadline={articleHeadline}
           onCopyLink={onCopyLink}
           onSaveToMyArticles={onSaveToMyArticles}
+          saveApi={saveApi}
         />
       );
     });
 
     it("save and share bar renders correctly", () => {
-      debugger;
       expect(testInstance.toJSON()).toMatchSnapshot();
-      expect(testInstance.root.findAllByType(BarItem)).toHaveLength(5);
+      expect(testInstance.root.findAllByType(BarItem)).toHaveLength(4);
     });
     it("onPress events triggers correctly", () => {
       testInstance.root.findAllByType(BarItem)[3].props.onPress();
       expect(Clipboard.setString).toHaveBeenCalledWith(articleUrl);
       expect(onCopyLink).toHaveBeenCalled();
-
-      testInstance.root.findAllByType(BarItem)[4].props.onPress();
-      expect(onSaveToMyArticles).toHaveBeenCalled();
     });
   });
 };