+++ packages/save-star-web/__tests__/shared.web.js
@@ -1,0 +1,24 @@
+import {
+  addSerializers,
+  compose,
+  minimaliseTransform,
+  minimalWebTransform,
+  print
+} from "@times-components/jest-serializer";
+import shared from "./shared.base";
+
+export default () => {
+  addSerializers(
+    expect,
+    compose(
+      print,
+      minimalWebTransform,
+      minimaliseTransform(
+        (value, key) =>
+          key === "style" || key === "className" || key === "data-testid"
+      )
+    )
+  );
+
+  shared();
+};