+++ packages/save-star-web/save-star-web.stories.js
@@ -1,0 +1,4 @@
+import { showcaseConverter } from "@times-components/storybook";
+import showcase from "./save-star-web.showcase";
+
+showcaseConverter(module, showcase);