+++ packages/utils/src/make-client-util.js
@@ -5,26 +5,26 @@ import { fragmentMatcher } from "@times-components/schema";
 import { InMemoryCache } from "apollo-cache-inmemory";
 import { ApolloClient } from "apollo-client";
 
- const makeClient = () => {
+const makeClient = () => {
   let graphqlapi = null;
   let acsTnlCookie = null;
   let sacsTnlCookie = null;
 
-   if (typeof window !== "undefined" && window.nuk) {
+  if (typeof window !== "undefined" && window.nuk) {
     graphqlapi = window.nuk.graphqlapi.url;
     acsTnlCookie = window.nuk.getCookieValue("acs_tnl");
     sacsTnlCookie = window.nuk.getCookieValue("sacs_tnl");
   }
   const networkInterfaceOptions = { fetch, headers: {}, uri: graphqlapi };
 
-   networkInterfaceOptions.headers["content-type"] =
+  networkInterfaceOptions.headers["content-type"] =
     "application/x-www-form-urlencoded";
   networkInterfaceOptions.headers.Authorization = `Cookie acs_tnl=${acsTnlCookie};sacs_tnl=${sacsTnlCookie}`;
 
-   return new ApolloClient({
+  return new ApolloClient({
     cache: new InMemoryCache({ fragmentMatcher }),
     link: createHttpLink(networkInterfaceOptions)
   });
 };
 
- export default makeClient;
+export default makeClient;