+++ ios-app/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.7.142](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.141...@times-components/ios-app@0.7.142) (2019-05-15)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.141](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.140...@times-components/ios-app@0.7.141) (2019-05-15)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.140](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.139...@times-components/ios-app@0.7.140) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
+## [0.7.139](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.138...@times-components/ios-app@0.7.139) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/ios-app
+
+
+
+
+
 ## [0.7.138](https://github.com/newsuk/times-components/compare/@times-components/ios-app@0.7.137...@times-components/ios-app@0.7.138) (2019-05-14)
 
 **Note:** Version bump only for package @times-components/ios-app