+++ packages/article-extras/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.32](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.31...@times-components/article-extras@0.1.32) (2019-05-15)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.1.31](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.30...@times-components/article-extras@0.1.31) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
+## [0.1.30](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.29...@times-components/article-extras@0.1.30) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/article-extras
+
+
+
+
+
 ## [0.1.29](https://github.com/newsuk/times-components/compare/@times-components/article-extras@0.1.28...@times-components/article-extras@0.1.29) (2019-05-14)
 
 **Note:** Version bump only for package @times-components/article-extras