+++ packages/key-facts/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.0.9](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.0.8...@times-components/key-facts@2.0.9) (2019-05-14)
+
+**Note:** Version bump only for package @times-components/key-facts
+
+
+
+
+
 ## [2.0.8](https://github.com/newsuk/times-components/compare/@times-components/key-facts@2.0.7...@times-components/key-facts@2.0.8) (2019-05-09)
 
 **Note:** Version bump only for package @times-components/key-facts